通信先の同時アクセス制御・ポリシー制御をResolver内のHTTPリクエスト毎に適用する
