+++ app/MetadataResolver/MetadataResolver.php
@@ -46,13 +46,17 @@ class MetadataResolver implements Resolver
 
     public $defaultResolver = OGPResolver::class;
 
+    public function __construct(private Client $client)
+    {
+    }
+
     public function resolve(string $url): Metadata
     {
         foreach ($this->rules as $pattern => $class) {
             if (preg_match($pattern, $url) === 1) {
                 try {
                     /** @var Resolver $resolver */
-                    $resolver = app($class);
+                    $resolver = app($class, ['client' => $this->client]);
 
                     return $resolver->resolve($url);
                 } catch (UnsupportedContentException $e) {
@@ -67,7 +71,7 @@ class MetadataResolver implements Resolver
 
         if (isset($this->defaultResolver)) {
             /** @var Resolver $resolver */
-            $resolver = app($this->defaultResolver);
+            $resolver = app($this->defaultResolver, ['client' => $this->client]);
 
             return $resolver->resolve($url);
         }
@@ -84,8 +88,7 @@ class MetadataResolver implements Resolver
             // Acceptヘッダには */* を足さないことにする。
             $acceptTypes = array_diff(array_keys($this->mimeTypes), ['*/*']);
 
-            $client = app(Client::class);
-            $res = $client->request('GET', $url, [
+            $res = $this->client->request('GET', $url, [
                 'headers' => [
                     'Accept' => implode(', ', $acceptTypes)
                 ]
@@ -97,14 +100,14 @@ class MetadataResolver implements Resolver
 
                 if (isset($this->mimeTypes[$mimeType])) {
                     $class = $this->mimeTypes[$mimeType];
-                    $parser = app($class);
+                    $parser = app($class, ['client' => $this->client]);
 
                     return $parser->parse($res->getBody());
                 }
 
                 if (isset($this->mimeTypes['*/*'])) {
                     $class = $this->mimeTypes['*/*'];
-                    $parser = app($class);
+                    $parser = app($class, ['client' => $this->client]);
 
                     return $parser->parse($res->getBody());
                 }
