+++ storage/global_lock/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
